fileinfo: move struct and default method into file_info file
