+++ agent/src/file_info.rs
