+++ agent/src/http_server/routes/get_heaviest_files.rs
@@ -1,4 +1,4 @@
-use crate::tidyalgo::FileInfo;
+use crate::file_info::FileInfo;
 use axum::Json;
 
 pub async fn get_heaviest_files() -> Json<Vec<FileInfo>> {
