--- agent/src/tidyalgo/mod.rs
@@ -1,3 +1,0 @@
-pub mod tidyalgo;
-pub use tidyalgo::FileInfo;
-pub use tidyalgo::TidyScore;
